+++ super_editor_spellcheck/lib/super_editor_spellcheck.dart
@@ -1,2 +1,3 @@
 export 'src/messages.g.dart';
 export 'src/spell_checker.dart';
+export 'src/spell_checker_mac.dart';
