+++ super_editor_spellcheck/pigeon/messages.dart
@@ -6,37 +6,41 @@ import 'package:pigeon/pigeon.dart';
 ))
 @HostApi()
 abstract class SpellCheckMac {
-  /// A list containing all the available spell checking languages. The languages are ordered
-  /// in the user’s preferred order as set in the system preferences.
+  /// {@template mac_spell_checker_available_languages}
+  /// A list containing all the available spell checking languages.
+  ///
+  /// The languages are ordered in the user’s preferred order as set in the
+  /// system preferences.
+  /// {@endtemplate}
   List<String?> availableLanguages();
 
-  /// Returns a unique tag to identified this spell checked object.
+  /// {@template mac_spell_checker_unique_spell_document_tag}
+  /// Returns a unique tag that identifies a single document in the spell checking system.
   ///
-  /// Use this method to generate tags to avoid collisions with other objects that can be spell checked.
+  /// Use this method to generate tags to avoid collisions when there are multiple different
+  /// texts being spell checked.
+  /// {@endtemplate}
   int uniqueSpellDocumentTag();
 
-  /// Notifies the receiver that the user has finished with the tagged document.
+  /// {@template mac_spell_checker_close_spell_document}
+  /// Notifies the spell checking system that the user has finished with the tagged document.
   ///
   /// The spell checker will release any resources associated with the document,
   /// including but not necessarily limited to, ignored words.
+  /// {@endtemplate}
   void closeSpellDocument(int tag);
 
-  /// Searches for a misspelled word in [stringToCheck] starting at [startingOffset]
-  /// within the string object.
+  /// {@template mac_spell_checker_check_spelling}
+  /// Searches for a misspelled word in [stringToCheck], starting at [startingOffset], and returns the
+  /// [TextRange] surrounding the misspelled word.
   ///
-  /// - [stringToCheck]: The string object containing the words to spellcheck.
-  /// - [startingOffset]: The offset within the string object at which to start the spellchecking.
-  /// - [language]: The language of the words in the string.
-  /// - [wrap]: `true` to indicate that spell checking should continue at the beginning of the string
-  ///   when the end of the string is reached; `false` to indicate that spellchecking should stop
-  ///   at the end of the string.
-  /// - [inSpellDocumentWithTag]: An identifier unique within the application
-  ///   used to inform the spell checker which document that text is associated, potentially
-  ///   for many purposes, not necessarily just for ignored words. A value of 0 can be passed
-  ///   in for text not associated with a particular document.
+  /// If no misspelled word is found, a [TextRange] is returned with bounds of `-1`, which can also be
+  /// queried more conveniently with [TextRange.isValid].
   ///
-  /// Returns the range of the first misspelled word.
-  Range checkSpelling({
+  /// To find all (or multiple) misspelled words in a given string, call this
+  /// method repeatedly, passing in different values for [startingOffset].
+  /// {@endtemplate}
+  PigeonRange checkSpelling({
     required String stringToCheck,
     required int startingOffset,
     String? language,
@@ -44,36 +48,34 @@ abstract class SpellCheckMac {
     int inSpellDocumentWithTag = 0,
   });
 
-  /// Returns an array of possible substitutions for the specified string.
+  /// {@template mac_spell_checker_guesses}
+  /// Returns possible substitutions for the specified misspelled word at [range] inside the [text].
   ///
-  /// - [range]: The range of the string to check.
-  /// - [text]: The string to guess.
-  /// - [language]: The language of the string.
-  /// - [inSpellDocumentWithTag]: An identifier unique within the application
-  ///   used to inform the spell checker which document that text is associated, potentially
-  ///   for many purposes, not necessarily just for ignored words. A value of 0 can be passed
-  ///   in for text not associated with a particular document.
-  ///
-  /// Returns an array of strings containing possible replacement words.
+  /// - [range]: The range, within the [text], for which possible substitutions should be generated.
+  /// - [text]: The string containing the word/text for which substitutions should be generated.
+  /// - [inSpellDocumentWithTag]: The (optional) ID of the loaded document that contains the given [text],
+  ///   which is used to provide additional context to the substitution guesses. A value of '0' instructs
+  ///   the guessing system to consider the [text] in isolation, without connection to any given document.
+  /// {@endtemplate}
   List<String?>? guesses({
     required String text,
-    required Range range,
+    required PigeonRange range,
     String? language,
     int inSpellDocumentWithTag = 0,
   });
 
-  /// Performs a grammatical analysis of a given string.
+  /// {@template mac_spell_checker_check_grammar}
+  /// Performs a grammatical analysis of [stringToCheck], starting at [startingOffset].
   ///
-  /// - [stringToCheck]: The string to analyze.
-  /// - [startingOffset]: Location within string at which to start the analysis.
-  /// - [language]: Language to use in string.
-  /// - [wrap]: `true` to specify that the analysis continue to the beginning of string when
-  ///   the end is reached. `false` to have the analysis stop at the end of string.
-  /// - [inSpellDocumentWithTag]: An identifier unique within the application
-  ///   used to inform the spell checker which document that text is associated, potentially
-  ///   for many purposes, not necessarily just for ignored words. A value of 0 can be passed
-  ///   in for text not associated with a particular document.
-  PlatformCheckGrammarResult checkGrammar({
+  /// - [stringToCheck]: The string containing the text to be analyzed.
+  /// - [startingOffset]: Location within the text at which the analysis should start.
+  /// - [wrap]: `true` to specify that the analysis continue to the beginning of the text when
+  ///   the end is reached. `false` to have the analysis stop at the end of the text.
+  /// - [inSpellDocumentWithTag]: The (optional) ID of the loaded document that contains the given [text],
+  ///   which is used to provide additional context to the substitution guesses. A value of '0' instructs
+  ///   the guessing system to consider the [stringToCheck] in isolation, without connection to any given document.
+  /// {@endtemplate}
+  PigeonCheckGrammarResult checkGrammar({
     required String stringToCheck,
     required int startingOffset,
     String? language,
@@ -81,50 +83,70 @@ abstract class SpellCheckMac {
     int inSpellDocumentWithTag = 0,
   });
 
-  /// Provides a list of complete words that the user might be trying to type based on a
-  /// partial word in a given string.
+  /// {@template mac_spell_checker_completions}
+  /// Provides a list of complete words that the user might be trying to type based on a partial word
+  /// at [partialWordRange] in the given [text].
   ///
-  /// - [partialWordRange] - Range that identifies a partial word in string.
-  /// - [text] - String with the partial word from which to generate the result.
-  /// - [language]: Language to use in string.
-  /// - [inSpellDocumentWithTag]: An identifier unique within the application
-  ///   used to inform the spell checker which document that text is associated, potentially
-  ///   for many purposes, not necessarily just for ignored words. A value of 0 can be passed
-  ///   in for text not associated with a particular document.
+  /// - [partialWordRange] - The range, within the [text], for which possible completions should be generated.
+  /// - [text] - The string containing the partial word for which completions should be generated.
+  /// - [inSpellDocumentWithTag]: The (optional) ID of the loaded document that contains the given [text],
+  ///   which is used to provide additional context to the substitution guesses. A value of '0' instructs
+  ///   the guessing system to consider the [text] in isolation, without connection to any given document.
   ///
-  /// Returns the list of complete words from the spell checker dictionary in the order
-  /// they should be presented to the user.
+  /// The items of the list are in the order they should be presented to the user.
+  /// {@endtemplate}
   List<String>? completions({
-    required Range partialWordRange,
+    required PigeonRange partialWordRange,
     required String text,
     String? language,
     int inSpellDocumentWithTag = 0,
   });
 
+  /// {@template mac_spell_checker_count_words}
   /// Returns the number of words in the specified string.
+  /// {@endtemplate}
   int countWords({required String text, String? language});
 
+  /// {@template mac_spell_checker_learn_word}
   /// Adds the [word] to the spell checker dictionary.
+  /// {@endtemplate}
   void learnWord(String word);
 
+  /// {@template mac_spell_checker_has_learned_word}
   /// Indicates whether the spell checker has learned a given word.
+  /// {@endtemplate}
   bool hasLearnedWord(String word);
 
+  /// {@template mac_spell_checker_unlearn_word}
   /// Tells the spell checker to unlearn a given word.
+  /// {@endtemplate}
   void unlearnWord(String word);
 
+  /// {@template mac_spell_checker_ignore_word}
   /// Instructs the spell checker to ignore all future occurrences of [word] in the document
   /// identified by [documentTag].
+  /// {@endtemplate}
   void ignoreWord({required String word, required int documentTag});
 
+  /// {@template mac_spell_checker_ignored_words}
   /// Returns the array of ignored words for a document identified by [documentTag].
+  /// {@endtemplate}
   List<String>? ignoredWords(int documentTag);
 
+  /// {@template mac_spell_checker_set_ignored_words}
   /// Updates the ignored-words document (a dictionary identified by [documentTag] with [words])
   /// with a list of [words] to ignore.
+  /// {@endtemplate}
   void setIgnoredWords({required List<String> words, required int documentTag});
 
-  /// Returns the dictionary used when replacing words.
+  /// {@template mac_spell_checker_user_replacements_dictionary}
+  /// Returns the dictionary used when replacing words, as defined by the user in the system preferences.
+  ///
+  /// This can be used to create an UI with replacement options when the user types a certain
+  /// combination of characters. For example, the user might want to automatically replace
+  /// "omw" with "on my way". When the user types "omw", an UI should display "on my way" as
+  /// a possible replacement.
+  /// {@endtemplate}
   Map<String, String> userReplacementsDictionary();
 }
 
@@ -134,8 +156,8 @@ abstract class SpellCheckMac {
 /// the one at [end].
 ///
 /// This is used because we can't use `TextRange` in pigeon.
-class Range {
-  Range({
+class PigeonRange {
+  PigeonRange({
     required this.start,
     required this.end,
   });
@@ -145,29 +167,29 @@ class Range {
 }
 
 /// The result of a grammatical analysis.
-class PlatformCheckGrammarResult {
-  PlatformCheckGrammarResult({
+class PigeonCheckGrammarResult {
+  PigeonCheckGrammarResult({
     this.firstError,
     this.details,
   });
 
   /// The range of the first error found in the text or `null` if no errors were found.
-  final Range? firstError;
+  final PigeonRange? firstError;
 
   /// A list of details about the grammatical errors found in the text or `null`
   /// if no errors were found.
-  final List<PlatformGrammaticalAnalysisDetail?>? details;
+  final List<PigeonGrammaticalAnalysisDetail?>? details;
 }
 
 /// A detail about a grammatical error found in a text.
-class PlatformGrammaticalAnalysisDetail {
-  PlatformGrammaticalAnalysisDetail({
+class PigeonGrammaticalAnalysisDetail {
+  PigeonGrammaticalAnalysisDetail({
     required this.range,
     required this.userDescription,
   });
 
   /// The range of the grammatical error in the text.
-  final Range range;
+  final PigeonRange range;
 
   /// A description of the grammatical error.
   final String userDescription;
